+++ azure-iot-device/azure/iot/device/iothub/transport/mqtt/mqtt_transport.py
@@ -9,10 +9,10 @@ from azure.iot.device.common.transport import pipeline_stages_base
 from azure.iot.device.common.transport import pipeline_ops_base
 from azure.iot.device.common.transport.mqtt import pipeline_stages_mqtt
 from azure.iot.device.iothub.transport.abstract_transport import AbstractTransport
+from azure.iot.device.iothub.transport import pipeline_stages_iothub
+from azure.iot.device.iothub.transport import pipeline_events_iothub
+from azure.iot.device.iothub.transport import pipeline_ops_iothub
 from . import pipeline_stages_iothub_mqtt
-from . import pipeline_stages_iothub
-from . import pipeline_ops_iothub
-from . import pipeline_events_iothub
 
 logger = logging.getLogger(__name__)
